music_player/requirements.txt
PyQt6
Pillow
mutagen
